requirements.txt
absl-py==2.1.0
accelerate==0.17.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.0
altair==5.0.1
annotated-types==0.5.0
antlr4-python3-runtime==4.8
anyio==3.7.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
backports.zoneinfo==0.2.1
basicsr==1.4.2
bitsandbytes==0.44.1
blinker==1.6.3
braceexpand==0.1.7
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cycler==0.11.0
decorator==5.1.1
einops==0.3.0
exceptiongroup==1.2.2
fastapi==0.103.2
ffmpy==0.3.2
filelock==3.12.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==0.4.6
gradio==3.16.2
grpcio==1.62.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.17.3
kiwisolver==1.4.5
kornia==0.6.0
linkify-it-py==2.0.3
lmdb==1.6.2
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mdit-py-plugins==0.3.5
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
omegaconf==2.1.1
open-clip-torch==2.0.2
opencv-contrib-python==4.3.0.36
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.9.7
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
prettytable==3.6.0
protobuf==3.20.3
psutil==7.0.0
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodome==3.22.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b1
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.17.2
Pympler==1.1
pyparsing==3.1.4
pyre-extensions==0.0.23
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
python-multipart==0.0.8
pytorch-lightning==1.5.0
pytz==2025.1
PyWavelets==1.3.0
PyYAML==6.0.1
qudida==0.0.4
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
semver==3.0.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.27.0
streamlit==1.12.1
streamlit-drawable-canvas==0.8.0
tb-nightly==2.12.0a20230113
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
test_tube==0.7.5
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.12
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==1.13.1+cu117
torchaudio==0.13.1+rocm5.2
torchmetrics==0.6.0
torchvision==0.14.1+cu117
tornado==6.2
tqdm==4.67.1
transformers==4.25.1
triton==2.3.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzlocal==5.1
uc-micro-py==1.0.3
urllib3==2.0.7
uvicorn==0.22.0
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.13
webdataset==0.2.5
websockets==11.0.3
Werkzeug==2.2.3
xformers==0.0.16
yapf==0.32.0
yarl==1.9.4
zipp==3.15.0
